requirements.txt
Flask==2.3.2
redis==4.5.4
boto3==1.34.39
Flask-WTF==1.2.1
python-dotenv==1.0.0
mysql-connector-python==8.2.0